ANOTHER attempt at STOP codon fitness
* This time we try assigning -inifity fitness to all stop codon sites at
any position in a feature to which a site-specific fitness
function (like purifying fitness) is applied.  Previously we only
considered sites within the fitness factor suring the initial assignment
of fitness to the genome, which meant that genomes that started off with
a STOP codon in a feature would be able to reproduce.

In other words, when initially assigning fitness values to genomes,
consider each site in a feature, not just the sites covered by a fitness
factor.  This is consistent with later treatment that considers every
site when updating the fitness of a feature.  If we assign fitness
values based on every site in the feature, then genomes with stop codons
will start out with -inifity fitness and will never be a candidate for
reproducion.  That way there is no opportunity for a site holding a STOP
to mutate away from the STOP, which avoids the whole issue
of generating NAN fitness values.